# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OwnChat API
  version: 1.0.0
  description: |
    Versioned API v1 for OwnChat behind FastAPI API Gateway.
    Base REST prefix: /api/v1
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local development
tags:
  - name: Auth
  - name: Chats
  - name: Messages
  - name: Media
  - name: Realtime
security:
  - bearerAuth: []
paths:
  /api/v1/auth/request-otp:
    post:
      tags: [Auth]
      summary: Request OTP for phone authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOtpRequest'
      responses:
        '200':
          description: OTP request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestOtpResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  /api/v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOtpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register user after OTP verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with verified OTP flow and get tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/chats:
    get:
      tags: [Chats]
      summary: Get user chats with cursor pagination
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Paginated chats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Chats]
      summary: Create a new chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
      responses:
        '201':
          description: Chat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/chats/{chat_id}/messages:
    get:
      tags: [Messages]
      summary: Get chat messages with bidirectional cursor pagination
      parameters:
        - $ref: '#/components/parameters/ChatIdParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: direction
          in: query
          required: false
          schema:
            type: string
            enum: [older, newer]
            default: older
      responses:
        '200':
          description: Paginated messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Messages]
      summary: Send a message
      parameters:
        - $ref: '#/components/parameters/ChatIdParam'
        - name: Idempotency-Key
          in: header
          required: false
          description: Idempotency key for safe retries
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '202':
          description: Message accepted for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/chats/{chat_id}/messages/read:
    post:
      tags: [Messages]
      summary: Mark message as read (REST fallback for read receipt)
      parameters:
        - $ref: '#/components/parameters/ChatIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadMessageRequest'
      responses:
        '200':
          description: Read status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadMessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/media/uploads:
    post:
      tags: [Media]
      summary: Create pre-signed upload URL for image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadRequest'
      responses:
        '201':
          description: Upload target created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /ws/v1/connect:
    get:
      tags: [Realtime]
      summary: WebSocket handshake endpoint
      description: |
        Upgrade endpoint for realtime events.
        Client authenticates by JWT query parameter `token` or by bearer token during gateway handshake.

        Events:
        - `message.new`
        - `typing.start`
        - `typing.stop`
        - `typing.update`
        - `message.read`
        - `message.read.updated`
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocols (WebSocket upgrade)
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ChatIdParam:
      name: chat_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
          example: Invalid token

    RequestOtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          example: '+79991234567'

    RequestOtpResponse:
      type: object
      required: [otp_session_id, expires_in]
      properties:
        otp_session_id:
          type: string
          format: uuid
        expires_in:
          type: integer
          example: 300

    VerifyOtpRequest:
      type: object
      required: [otp_session_id, code]
      properties:
        otp_session_id:
          type: string
          format: uuid
        code:
          type: string
          minLength: 4
          maxLength: 8
          example: '123456'

    VerifyOtpResponse:
      type: object
      required: [verified, verification_token]
      properties:
        verified:
          type: boolean
          example: true
        verification_token:
          type: string
          description: Short-lived token proving 2FA completion

    RegisterRequest:
      type: object
      required: [verification_token, display_name]
      properties:
        verification_token:
          type: string
        display_name:
          type: string
          minLength: 1
          maxLength: 64
        avatar_url:
          type: string
          format: uri

    RegisterResponse:
      type: object
      required: [user_id, created_at]
      properties:
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

    LoginRequest:
      type: object
      required: [verification_token]
      properties:
        verification_token:
          type: string
        device_id:
          type: string

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    TokenPairResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900

    MessageType:
      type: string
      enum: [text, image]

    DeliveryStatus:
      type: string
      enum: [sent, delivered, read]

    LastMessage:
      type: object
      required: [id, type, created_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/MessageType'
        content_preview:
          type: string
        created_at:
          type: string
          format: date-time

    ChatItem:
      type: object
      required: [chat_id, unread_count, updated_at]
      properties:
        chat_id:
          type: string
          format: uuid
        title:
          type: string
        last_message:
          $ref: '#/components/schemas/LastMessage'
        unread_count:
          type: integer
          minimum: 0
        updated_at:
          type: string
          format: date-time

    ChatListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChatItem'
        next_cursor:
          type: string
          nullable: true

    CreateChatRequest:
      type: object
      required: [participant_ids]
      properties:
        participant_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        title:
          type: string

    CreateChatResponse:
      type: object
      required: [chat_id, created_at]
      properties:
        chat_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

    Message:
      type: object
      required: [id, chat_id, author_id, type, created_at, status]
      properties:
        id:
          type: string
          format: uuid
        chat_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/MessageType'
        content:
          type: string
          nullable: true
        attachment_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/DeliveryStatus'

    MessageHistoryResponse:
      type: object
      required: [items, next_cursor, prev_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true

    SendTextMessageRequest:
      type: object
      required: [type, content, client_message_id]
      properties:
        type:
          type: string
          enum: [text]
        content:
          type: string
          minLength: 1
        client_message_id:
          type: string
          format: uuid

    SendImageMessageRequest:
      type: object
      required: [type, attachment_id, client_message_id]
      properties:
        type:
          type: string
          enum: [image]
        attachment_id:
          type: string
          format: uuid
        client_message_id:
          type: string
          format: uuid

    SendMessageRequest:
      oneOf:
        - $ref: '#/components/schemas/SendTextMessageRequest'
        - $ref: '#/components/schemas/SendImageMessageRequest'

    SendMessageResponse:
      type: object
      required: [message_id, status, created_at]
      properties:
        message_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/DeliveryStatus'
        created_at:
          type: string
          format: date-time

    ReadMessageRequest:
      type: object
      required: [message_id, read_at]
      properties:
        message_id:
          type: string
          format: uuid
        read_at:
          type: string
          format: date-time

    ReadMessageResponse:
      type: object
      required: [chat_id, message_id, user_id, status, read_at]
      properties:
        chat_id:
          type: string
          format: uuid
        message_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [read]
        read_at:
          type: string
          format: date-time

    CreateUploadRequest:
      type: object
      required: [content_type, file_name, size_bytes]
      properties:
        content_type:
          type: string
          example: image/jpeg
        file_name:
          type: string
          example: photo.jpg
        size_bytes:
          type: integer
          minimum: 1

    CreateUploadResponse:
      type: object
      required: [attachment_id, upload_url, expires_in]
      properties:
        attachment_id:
          type: string
          format: uuid
        upload_url:
          type: string
          format: uri
        expires_in:
          type: integer
          example: 300